[Add]Timerコンポーネントに移行
